pyproject.toml
[project]
name = "apibot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot[job-queue,webhooks]>=21.6",
    "requests>=2.32.3",
]
